event (price) endpoint SQL integration tests

+++ backend/src/routes/event_price/handlers/fetch.ts
@@ -19,7 +19,7 @@ export async function fetchEventPrice(
 ): Promise<EventPrice | null> {
   const { id } = parseParams(request, z.object({ id: stringToInt }));
   const result = await buildQuery(server,
-    `SELECT id, event, amount
+    `SELECT id, event, amount::float
     FROM event_price WHERE id = $1`,
     z.tuple([z.int()]),
     EventPriceSchema,
@@ -43,7 +43,7 @@ export async function fetchEventPriceWithMeta(
   const { id } = parseParams(request, z.object({ id: stringToInt }));
   const result = await buildQuery(
     server,
-    `SELECT id, event, amount, created_at, updated_at, created_by, updated_by
+    `SELECT id, event, amount::float, created_at, updated_at, created_by, updated_by
     FROM event_price WHERE id = $1`,
     z.tuple([z.int()]),
     EventPriceSchema.withMeta(),
@@ -66,7 +66,7 @@ export async function fetchEventPrices(
 ): Promise<EventPrice[]> {
   const result = await buildQuery(
     server,
-    `SELECT id, event, amount
+    `SELECT id, event, amount::float
     FROM event_price`,
     EventPriceSchema,
   )();
